+++ klibc/string/k_strncmp.c
@@ -4,7 +4,7 @@
 
 #include "string.h"
 
-int k_strncmp(const char *s1, const char *s2, size_t n) {
+int k_strncmp(const char *s1, const char *s2, uint32_t n) {
     const unsigned char *p1 = (const unsigned char*) s1;
     const unsigned char *p2 = (const unsigned char*) s2;
     unsigned char	u1;
